Initial commit Delete unnecessary files and restructured solution
