+++ ObjectPrinting/ObjectExtensions.cs
@@ -1,4 +1,4 @@
-namespace ObjectPrinting.Solved
+namespace ObjectPrinting
 {
     public static class ObjectExtensions
     {
