--- ObjectPrinting/Solved/PrintingConfig.cs
@@ -1,62 +1,0 @@
-﻿using System;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Text;
-
-namespace ObjectPrinting.Solved
-{
-    public class PrintingConfig<TOwner>
-    {
-        public PropertyPrintingConfig<TOwner, TPropType> Printing<TPropType>()
-        {
-            return new PropertyPrintingConfig<TOwner, TPropType>(this);
-        }
-
-        public PropertyPrintingConfig<TOwner, TPropType> Printing<TPropType>(Expression<Func<TOwner, TPropType>> memberSelector)
-        {
-            return new PropertyPrintingConfig<TOwner, TPropType>(this);
-        }
-
-        public PrintingConfig<TOwner> Excluding<TPropType>(Expression<Func<TOwner, TPropType>> memberSelector)
-        {
-            return this;
-        }
-
-        internal PrintingConfig<TOwner> Excluding<TPropType>()
-        {
-            return this;
-        }
-
-        public string PrintToString(TOwner obj)
-        {
-            return PrintToString(obj, 0);
-        }
-
-        private string PrintToString(object obj, int nestingLevel)
-        {
-            //TODO apply configurations
-            if (obj == null)
-                return "null" + Environment.NewLine;
-
-            var finalTypes = new[]
-            {
-                typeof(int), typeof(double), typeof(float), typeof(string),
-                typeof(DateTime), typeof(TimeSpan)
-            };
-            if (finalTypes.Contains(obj.GetType()))
-                return obj + Environment.NewLine;
-
-            var identation = new string('\t', nestingLevel + 1);
-            var sb = new StringBuilder();
-            var type = obj.GetType();
-            sb.AppendLine(type.Name);
-            foreach (var propertyInfo in type.GetProperties())
-            {
-                sb.Append(identation + propertyInfo.Name + " = " +
-                          PrintToString(propertyInfo.GetValue(obj),
-                              nestingLevel + 1));
-            }
-            return sb.ToString();
-        }
-    }
-}
