--- ObjectPrinting/Tests/Person.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace ObjectPrinting.Tests
-{
-    public class Person
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; }
-        public double Height { get; set; }
-        public int Age { get; set; }
-    }
-}
